+++ source/core/custom/template/command.js
@@ -6,7 +6,8 @@ module.exports = () => ({
   exec: {
     _: 'M',
     'z1': '#{bin}',
-    'z2': '#{cmd}'
+    'z2': '#{cmd}',
+    'z3': '#{env}'
   },
   listcmd: {
     _: 'Y',